requirements.txt
pillow==10.2.0
natsort==8.4.0